refactor: extract natural_sort_key to shared utils and fix test pattern
Addresses review comments:
- Extract duplicated natural_sort_key to generators/utils.py
- Make test pattern more specific for interface address commands
